Add order entry and move dto folder
New class will allow specify quantity of product instead of duplicating
products in list.
