+++ src/main/java/com/example/cosmocats/domain/order/OrderEntry.java
@@ -1,0 +1,8 @@
+package com.example.cosmocats.domain.order;
+
+import com.example.cosmocats.domain.Product;
+
+public class OrderEntry {
+    Product product;
+    int quantity;
+}
