+++ src/main/java/com/example/cosmocats/dto/order/OrderDto.java
@@ -1,4 +1,4 @@
-package com.example.dto;
+package com.example.cosmocats.dto.order;
 
 import lombok.Builder;
 import lombok.Value;
@@ -7,6 +7,6 @@ import lombok.Value;
 @Builder
 public class OrderDto {
     long id;
-    long[] productIds;
+    long[] entryIds;
     float price;
 }
