+++ src/main/java/com/example/cosmocats/dto/order/OrderEntryDto.java
@@ -1,0 +1,6 @@
+package com.example.cosmocats.dto.order;
+
+public class OrderEntryDto {
+    long product;
+    int quantity;
+}
